requirements.txt

python-dotenv
selenium==3.141.0
webdriver-manager

